feat: 남은 시간 계산 로직 분리
