+++ src/components/organisms/AuctionTimer/index.tsx
@@ -1,7 +1,6 @@
-import { useEffect, useState, useRef } from "react";
 import { Text } from "components/atoms";
-import { getTimeRemaining } from "utils";
 import { AuctionTimerWrapper } from "./styled";
+import { useRemainingTimer } from "hooks";
 
 interface IAuctionTimerProps {
   /** 경매 종료 시간 */
@@ -9,31 +8,7 @@ interface IAuctionTimerProps {
 }
 
 export const AuctionTimer = ({ targetDate }: IAuctionTimerProps) => {
-  const [timeRemaining, setTimeRemaining] = useState<string>("");
-  const intervalId = useRef<NodeJS.Timeout | null>(null);
-
-  useEffect(() => {
-    const updateTimer = () => {
-      const remainingTime = getTimeRemaining(targetDate);
-      setTimeRemaining(remainingTime);
-
-      if (intervalId.current && remainingTime === "over") {
-        clearInterval(intervalId.current);
-      }
-    };
-
-    updateTimer();
-
-    if (timeRemaining !== "over") {
-      intervalId.current = setInterval(updateTimer, 1000);
-    }
-
-    return () => {
-      if (intervalId.current) {
-        clearInterval(intervalId.current);
-      }
-    };
-  }, [targetDate, timeRemaining]);
+  const { timeRemaining } = useRemainingTimer(targetDate);
 
   return (
     <AuctionTimerWrapper>
@@ -47,4 +22,3 @@ export const AuctionTimer = ({ targetDate }: IAuctionTimerProps) => {
     </AuctionTimerWrapper>
   );
 };
-
